requirements.txt
async-timeout==4.0.2
discord.py==2.1.1
ffmpeg==1.4
yt-dlp==2023.2.17
PyNaCl==1.5.0